+++ lib/vagrant/guest.rb
@@ -5,11 +5,11 @@ module Vagrant
     # Specific guests
     autoload :Arch,    'vagrant/guest/arch'
     autoload :Debian,  'vagrant/guest/debian'
+    autoload :Fedora,  'vagrant/guest/fedora'
     autoload :FreeBSD, 'vagrant/guest/freebsd'
     autoload :Gentoo,  'vagrant/guest/gentoo'
     autoload :Linux,   'vagrant/guest/linux'
     autoload :Redhat,  'vagrant/guest/redhat'
-    autoload :Fedora,  'vagrant/guest/fedora'
     autoload :Solaris, 'vagrant/guest/solaris'
     autoload :Suse,    'vagrant/guest/suse'
     autoload :Ubuntu,  'vagrant/guest/ubuntu'