+++ CHANGELOG.md
@@ -19,6 +19,7 @@
     [GH-726]
   - Fix issue where Vagrant would sometimes "lose" a VM if an exception
     occurred. [GH-725]
+  - `vagrant destroy` destroys virtual machines in reverse order. [GH-739]
 
 ## 0.9.7 (February 9, 2012)