+++ CHANGELOG.md
@@ -22,6 +22,7 @@
   - `vagrant destroy` destroys virtual machines in reverse order. [GH-739]
   - Add an `fsid` option to Linux NFS exports. [GH-736]
   - Fix edge case where an exception could be raised in networking code. [GH-742]
+  - Add missing translation for the "guru meditation" state. [GH-745]
 
 ## 0.9.7 (February 9, 2012)