+++ CHANGELOG.md
@@ -24,7 +24,7 @@
   - Fix edge case where an exception could be raised in networking code. [GH-742]
   - Add missing translation for the "guru meditation" state. [GH-745]
   - Check that VirtualBox exists before certain commands. [GH-746]
-  - NIC type can be defined for host-only network adapters.
+  - NIC type can be defined for host-only network adapters. [GH-750]
 
 ## 0.9.7 (February 9, 2012)