+++ CHANGELOG.md
@@ -28,6 +28,7 @@
   - Fix issue where re-running chef-client would sometimes cause
     problems due to file permissions. [GH-748]
   - FreeBSD guests can now have their hostnames changed. [GH-757]
+  - FreeBSD guests now support host only networking and bridged networking. [GH-762]
 
 ## 0.9.7 (February 9, 2012)