+++ CHANGELOG.md
@@ -29,6 +29,7 @@
     problems due to file permissions. [GH-748]
   - FreeBSD guests can now have their hostnames changed. [GH-757]
   - FreeBSD guests now support host only networking and bridged networking. [GH-762]
+  - `VM#run_action` is now public so plugin-devs can hook into it.
 
 ## 0.9.7 (February 9, 2012)